refactor(renderer): `electron-renderer/plugins` -> `electron-renderer/src`
