+++ packages/electron-renderer/src/index.ts
@@ -1,10 +1,10 @@
 import type { Plugin } from 'vite'
-import buildConfig from './plugins/build-config'
-import polyfillExports from './plugins/polyfill-exports'
+import buildConfig from './build-config'
+import polyfillExports from './polyfill-exports'
 import {
   type Options,
   default as useNodeJs,
-} from './plugins/use-node.js'
+} from './use-node.js'
 
 export default function renderer(options: Options = {}): Plugin[] {
   return [
